% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_key}
\alias{check_key}
\title{Check SHERPA/RoMEO API key}
\usage{
check_key(key = NULL)
}
\arguments{
\item{key}{[\code{character(1)}]\cr{}
a character string containing the API key or \code{NULL}
(see Details section on how to specify it)}
}
\value{
if found the character string of the key, \code{NULL} otherwise
}
\description{
The key can be either specified in various ways see the Details section.
}
\details{
There are several ways to provide your API key.
The best way to know about them is to refer to the vignette about
"Setting Up Your API key" accessible with the following command:
\code{vignette("setting_up_api_key", package = "rromeo")}.
You can also use \code{\link{rr_auth}} that will use the provided key to store it as
an environmental variable.
}
