% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_embargo}
\alias{parse_embargo}
\title{Parse embargo period from API return}
\usage{
parse_embargo(xml_source, type)
}
\arguments{
\item{xml_source}{a parsed xml document}

\item{type}{name of the embargo type must be in \code{"pre"}, \code{"post"}, and
\code{"pdf"}}
}
\value{
the embargo period as a string like \code{"12 months"}
}
\description{
This function provides an easy way to return the embargo period in the
different categories
}
\keyword{internal}
