% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_base.R
\name{rr_api_version}
\alias{rr_api_version}
\title{Return SHERPA/RoMEO API version}
\usage{
rr_api_version()
}
\description{
This function queries SHERPA/RoMEO and returns the version of the API.
}
\examples{
\dontrun{
rr_api_version()
}
}
