% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_base.R
\name{rr_auth}
\alias{rr_auth}
\title{Store provided API key into Environment Variable}
\usage{
rr_auth(key)
}
\arguments{
\item{key}{[\code{character(1)}]\cr{}
A string giving the API key to save into the environment}
}
\description{
This function stores the provided API key as argument in to an environment
variable \code{SHERPAROMEO_KEY} for further use by other \code{rromeo} functions.
}
\details{
For more information regarding API keys, please refer to dedicated vignette
with the following command
\code{vignette("setting_up_api_key", package = "rromeo")}
}
\examples{
rr_auth("Iq83AIL5bss")
}
