% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_check_api}
\alias{rr_check_api}
\title{Checks SHERPA/RoMEO API}
\usage{
rr_check_api()
}
\value{
Nothing if can reach, error otherwise
}
\description{
If reaching the RoMEO API this function returns an error
}
\examples{
rr_check_api()
}
