% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rr_GET}
\alias{rr_GET}
\title{rromeo internal GET function}
\usage{
rr_GET(...)
}
\arguments{
\item{...}{additional parameter to \code{\link[httr:GET]{httr::GET}}}
}
\description{
rromeo internal GET function
}
