% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_get_rest_pub}
\alias{rr_get_rest_pub}
\title{Get record from Publication ID}
\usage{
rr_get_rest_pub(id)
}
\arguments{
\item{id}{publication number}
}
\description{
Use SHERPA/RoMEO RESTful API to retrieve a specific publication number status
}
