% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_journal}
\alias{rr_journal}
\title{Journal data by ISSN}
\usage{
rr_journal(issn)
}
\arguments{
\item{issn}{A single journal ISSN}
}
\description{
Journal data by ISSN
}
\examples{

rr_journal("1947-6264")
}
