% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_journal_issn}
\alias{rr_journal_issn}
\title{Journal data by ISSN}
\usage{
rr_journal_issn(issn)
}
\arguments{
\item{issn}{A single journal ISSN}
}
\description{
Journal data by ISSN
}
\examples{

rr_journal_issn("1947-6264")
}
