% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_publisher}
\alias{rr_publisher}
\title{Get record from Publication ID}
\usage{
rr_publisher(given_id)
}
\arguments{
\item{given_id}{publication number}
}
\description{
Use SHERPA/RoMEO regular API to retrieve a specific publication number status
}
\examples{
rr_publisher(55)
}
