% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_rest_publisher}
\alias{rr_rest_publisher}
\title{Get record from Publichser ID}
\usage{
rr_rest_publisher(id)
}
\arguments{
\item{id}{publisher ID}
}
\description{
Use SHERPA/RoMEO RESTful API to retrieve a specific publisher status
}
