% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo-package.R
\docType{package}
\name{rromeo-package}
\alias{rromeo}
\alias{rromeo-package}
\title{rromeo: Access Publisher Copyright & Self-Archiving Policies via the 
  'SHERPA/RoMEO' API}
\description{
Fetches information from the 'SHERPA/RoMEO' API
  <http://www.sherpa.ac.uk/romeo/apimanual.php> which indexes policies of
  journal regarding the archival of scientific manuscripts before and/or after
  peer-review as well as formatted manuscripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/rromeo/}
  \item \url{https://github.com/ropensci/rromeo}
  \item Report bugs at \url{https://github.com/ropensci/rromeo/issues}
}

}
\author{
\strong{Maintainer}: Matthias Grenié \email{matthias.grenie@gmail.com} (\href{https://orcid.org/0000-0002-4659-7522}{ORCID})

Authors:
\itemize{
  \item Hugo Gruson (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})
}

}
\keyword{internal}
