% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_text_or_na}
\alias{xml_text_or_na}
\title{Search xml tag and get character value}
\usage{
xml_text_or_na(parsed_xml, xml_tag)
}
\arguments{
\item{parsed_xml}{parsed xml document}

\item{xml_tag}{a string containing an xpath expression}
}
\value{
\code{NA} if the tag isn't found, the string of the tag content otherwise
}
\description{
This function extract the character value from a given parsed xml document
using \code{\link[xml2:xml_find_all]{xml2::xml_find_all()}} and \code{\link[xml2:xml_text]{xml2::xml_text()}}.
}
