% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_key}
\alias{check_key}
\title{Check SHERPA/RoMEO API key}
\usage{
check_key(key)
}
\arguments{
\item{key}{a character string or \code{NULL}}
}
\description{
The key can be either specified in an \code{.Renviron} file using
\code{SHERPAROMEO_KEY=my_key} or by passing it when calling the function.
}
