% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_key}
\alias{check_key}
\title{Check SHERPA/RoMEO API key}
\usage{
check_key(key = NULL)
}
\arguments{
\item{key}{{\code{character(1)}}\cr{}
a character string containing the API key or \code{NULL}
(see Details section on how to specify it)}
}
\value{
if found the character string of the key, \code{NULL} otherwise
}
\description{
The key can be either specified in various ways see the Details section.
}
\details{
To provide your API key to \code{rromeo} you can:
\enumerate{
\item provide it as a character string as the \code{key} arguments of \code{rromeo}
functions as \code{rr_*(..., key = "my_key_as_a_string")} (\strong{NOT RECOMMENDED})
\item you can define the variable \code{SHERPAROMEO_KEY} in an \code{.Renviron} file in
your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY=my_key_without_quotes}
\item you can also define the variable \code{SHERPAROMEO_KEY} in an \code{.Rprofile} file
in your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY="my_key_with_quotes"}
}
}
