% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_journal}
\alias{parse_journal}
\title{Parse API answer}
\usage{
parse_journal(xml_source, outcome, hits, multiple = FALSE, key = NULL)
}
\arguments{
\item{multiple}{{\code{logical(1)}}\cr{}
If multiple results match your query, should the function
recursively fetch data for each of one of them
(\code{multiple = TRUE}) or return a data.frame containing only
titles and ISSN of all matches (\code{multiple = FALSE})}

\item{key}{{\code{character(1)}}\cr{}
a character string containing the API key or \code{NULL}
(see Details section on how to specify it)}
}
\value{
Returns a data.frame with the following columns:
\itemize{
\item \code{title}        {\code{character(1)}}\cr{}
the name of the journal
\item \code{issn}         {\code{character(1)}}\cr{}
the ISSN of the journal
\item \code{romeocolour}  {\code{character(1)}}\cr{}
the SHERPA/RoMEO colour of the journal
\item \code{preprint}     {\code{character(1)}}\cr{}
is the preprint (not reviewed) archivable?
\item \code{postprint}    {\code{character(1)}}\cr{}
is the postprint (reviewed but not formatted) archivable?
\item \code{pdf}          {\code{character(1)}}\cr{}
is the publisher's version (reviewed and formatted)
\item \code{pre_embargo}  {\code{character(1)}}\cr{}
if applicable the embargo period before the author(s) can
archive the preprint
\item \code{post_embargo} {\code{character(1)}}\cr{}
if applicable the embargo period before the author(s) can
archive the postprint
\item \code{pdf_embargo}  {\code{character(1)}}\cr{}
if applicable the embargo period before the author(s) can
archive the publisher's version
}
}
\description{
Returns data.frame from parsed xml API answer.
}
\keyword{internal}
