% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_publisher}
\alias{parse_publisher}
\title{Parse publisher list}
\usage{
parse_publisher(api_answer)
}
\arguments{
\item{api_answer}{parsed xml API answer}
}
\description{
Parse publisher list
}
\keyword{internal}
