% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_journal_issn}
\alias{rr_journal_issn}
\title{Retrieve journal policy using ISSN}
\usage{
rr_journal_issn(issn, key = NULL)
}
\arguments{
\item{issn}{A single journal ISSN}

\item{key}{a character string or \code{NULL}}
}
\description{
Retrieve policy information from the SHERPA/RoMEO API using the ISSN of
the journal
}
\examples{

rr_journal_issn("1947-6264")
}
