% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_journal_issn}
\alias{rr_journal_issn}
\title{Retrieve journal policy using ISSN}
\usage{
rr_journal_issn(issn, key = NULL)
}
\arguments{
\item{issn}{\code{[character(1+)]} one or a vector of journal(s) ISSN(s)}

\item{key}{a character string containing the API key or \code{NULL}}
}
\description{
Retrieve policy information from the SHERPA/RoMEO API using the ISSN of
the journal
}
\examples{
rr_journal_issn("1947-6264")
rr_journal_issn(c("1947-6264", "0030-1299"))
}
