% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_journal_name}
\alias{rr_journal_name}
\title{Journal data by title}
\usage{
rr_journal_name(name, multiple = FALSE, qtype = c("exact", "contains",
  "starts with"))
}
\arguments{
\item{name}{A character string, containing the (possibly) partial name of the
journal}

\item{multiple}{If multiple results match your query, should the function
recursively fetch data for each of one of them (\code{multiple = TRUE}) or
return a data.frame containing only titles and ISSN of all matches
(\code{multiple = FALSE})}

\item{qtype}{A character string saying whether you are lookin for \code{exact},
\code{contains} or \code{starts with} matches}
}
\description{
Journal data by title
}
\examples{
rr_journal_name("Journal of Geology")
rr_journal_name("Biogeography", multiple = FALSE, qtype = "contains")
rr_journal_name("Biogeography", multiple = TRUE, qtype = "contains")
}
