% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_journal_name}
\alias{rr_journal_name}
\title{Retrieve journal(s) policy(ies) by matching title}
\usage{
rr_journal_name(name, multiple = FALSE, qtype = c("exact", "contains",
  "starts with"), key = NULL)
}
\arguments{
\item{name}{\code{[character(1+)]} one or several strings to match the titles of
the journals}

\item{multiple}{If multiple results match your query, should the function
recursively fetch data for each of one of them (\code{multiple = TRUE}) or
return a data.frame containing only titles and ISSN of all matches
(\code{multiple = FALSE})}

\item{qtype}{\code{[character(1)]} in \code{c("exact", "contains", "starts with")} to
set match type for the \code{name} search string}

\item{key}{a character string containing the API key or \code{NULL}
(see Details section on how to specify it)}
}
\value{
Returns a data frame if multiple journals are found and
\code{multiple = FALSE}:
\itemize{
\item \code{title}        \code{[character(1)]} the name of the journal
\item \code{issn}         \code{[character(1)]} the ISSN of the journal
if a single journal is found or if \code{multiple = TRUE} returns a larger
data frame:
\item \code{title}        \code{[character(1)]} the name of the journal
\item \code{issn}         \code{[character(1)]} the ISSN of the journal
\item \code{romeocolour}  \code{[character(1)]} the SHERPA/RoMEO colour of the journal
\item \code{preprint}     \code{[character(1)]} is the preprint (not reviewed) archivable?
\item \code{postprint}    \code{[character(1)]} is the postprint (reviewed but not
typesetted) archivable?
\item \code{pdf}          \code{[character(1)]} is the publisher's version
(reviewed and typesetted)
\item \code{pre_embargo}  \code{[character(1)]} if applicable the embargo period before
the author(s) can archive the pre-print
\item \code{post_embargo} \code{[character(1)]} if applicable the embargo period before
the author(s) can archive the post-print
\item \code{pdf_embargo}  \code{[character(1)]} if applicable the embargo period before
the author(s) can archive the publisher's version
}
}
\description{
Retrieve journal(s) policy(ies) by matching title
}
\details{
To provide your API key to \code{rromeo} you can:
\enumerate{
\item provide it as a character string as the \code{key} arguments of \code{rromeo}
functions as \code{rr_*(..., key = "my_key_as_a_string")} (\strong{NOT RECOMMENDED})
\item you can define the variable \code{SHERPAROMEO_KEY} in an \code{.Renviron} file in
your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY=my_key_without_quotes}
\item you can also define the variable \code{SHERPAROMEO_KEY} in an \code{.Rprofile} file
in your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY="my_key_with_quotes"}
}
}
\examples{
rr_journal_name(name = "Journal of Geology")
rr_journal_name(name = "Biogeography", multiple = FALSE, qtype = "contains")
\dontrun{
rr_journal_name(name = "Biogeography", multiple = TRUE, qtype = "contains")
# You can also query multiple journals with exact titles in a single call
rr_journal_name(name = c("Journal of Biogeography", "PLoS ONE"),
                qtype = "exact")
}
}
