% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_publisher}
\alias{rr_publisher}
\title{Get Publisher Information from Publisher ID}
\usage{
rr_publisher(romeo_id, key = NULL)
}
\arguments{
\item{romeo_id}{\code{[integer(1+)]} one or a vector of SHERPA/RoMEO publisher's
ID}

\item{key}{a character string containing the API key or \code{NULL}
(see Details section on how to specify it)}
}
\value{
Returns a data frame with the following columns:
\itemize{
\item \code{romeoid}     \code{[integer(1)]} the internal index of the publisher in
the SHERPA/RoMEO database
\item \code{publisher}   \code{[character(1)]} the name of the publisher
\item \code{alias}       \code{[character(1)]} if applicable an alternative name of the
publisher or the name of the specific publishing branch
\item \code{romeocolour} \code{[character(1)]} a colour assigned by the database that
reflects the default policies of the publisher
\item \code{preprint}    \code{[character(1)]} is the preprint (not reviewed) archivable?
\item \code{postprint}   \code{[character(1)]} is the postprint (reviewed but not
typesetted) archivable?
\item \code{pdf}         \code{[character(1)]} is the publisher's version (reviewed and
typesetted) archivable?
}
}
\description{
Use SHERPA/RoMEO API to retrieve a specific publisher policies on manuscript
archivals
}
\details{
To provide your API key to \code{rromeo} you can 1) provide it as a
character string as the \code{key} arguments of \code{rromeo} functions
as \code{rr_*(..., key = "my_key_as_a_string")} (\strong{NOT RECOMMENDED})
2) you can define the variable \code{SHERPAROMEO_KEY} in an \code{.Renviron}
file in your working directory, the file should contain the
following line \code{SHERPAROMEO_KEY=my_key_without_quotes}
3) you can also define the variable \code{SHERPAROMEO_KEY} in an
\code{.Rprofile} file in your working directory, the file should contain
the following line \code{SHERPAROMEO_KEY="my_key_with_quotes"}.
}
\examples{
rr_publisher(romeo_id = 55)
rr_publisher(romeo_id = c(55, 735))
}
