% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_publishers.R
\name{rr_publisher_all}
\alias{rr_publisher_all}
\title{Get all Publisher Policies}
\usage{
rr_publisher_all(key = NULL)
}
\arguments{
\item{key}{\code{\link{character(1)}}\cr{}
a character string containing the API key or \code{NULL}
(see Details section on how to specify it)}
}
\value{
Returns a data frame with the following columns:
\itemize{
\item \code{romeoid}     \code{\link{integer(1)}}\cr{}
the internal index of the publisher in the SHERPA/RoMEO
database
\item \code{publisher}   \code{\link{character(1)}}\cr{}
the name of the publisher
\item \code{alias}       \code{\link{character(1)}}\cr{}
if applicable an alternative name of the publisher or the
name of the specific publishing branch
\item \code{romeocolour} \code{\link{character(1)}}\cr{}
a colour assigned by the database that reflects the default
policies of the publisher
\item \code{preprint}    \code{\link{character(1)}}\cr{}
is the preprint (not reviewed) archivable?
\item \code{postprint}   \code{\link{character(1)}}\cr{}
is the postprint (reviewed but not formatted) archivable?
\item \code{pdf}         \code{\link{character(1)}}\cr{}
is the publisher's version (reviewed and formatted)
archivable?
}
}
\description{
Retrieve all data on publishers policies from SHERPA/RoMEO.
}
\details{
To provide your API key to \code{rromeo} you can:
\enumerate{
\item provide it as a character string as the \code{key} arguments of \code{rromeo}
functions as \code{rr_*(..., key = "my_key_as_a_string")} (\strong{NOT RECOMMENDED})
\item you can define the variable \code{SHERPAROMEO_KEY} in an \code{.Renviron} file in
your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY=my_key_without_quotes}
\item you can also define the variable \code{SHERPAROMEO_KEY} in an \code{.Rprofile} file
in your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY="my_key_with_quotes"}
}
}
