% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_publishers.R
\name{rr_publisher_continent}
\alias{rr_publisher_continent}
\title{Get Publisher Policy by Publisher's Continent}
\usage{
rr_publisher_continent(continent = c("Africa", "Antarctica", "Asia",
  "Australasia", "Caribbean", "Central America", "Europe", "North America",
  "Oceania", "South America"), key = NULL)
}
\arguments{
\item{continent}{{\code{character(1)}}\cr{}
in \code{c("Africa", "Antarctica",  "Asia",  "Australasia", "Carribean",  "Central America",  "Europe", "North America",  "Oceania",  "South America")}\cr{}
the continent name to retrieve}

\item{key}{{\code{character(1)}}\cr{}
a character string containing the API key or \code{NULL}
(see Details section on how to specify it)}
}
\value{
Returns a data frame with the following columns:
\itemize{
\item \code{romeoid}     {\code{integer(1)}}\cr{}
the internal index of the publisher in the SHERPA/RoMEO
database
\item \code{publisher}   {\code{character(1)}}\cr{}
the name of the publisher
\item \code{alias}       {\code{character(1)}}\cr{}
if applicable an alternative name of the publisher or the
name of the specific publishing branch
\item \code{romeocolour} {\code{character(1)}}\cr{}
a colour assigned by the database that reflects the default
policies of the publisher
\item \code{preprint}    {\code{character(1)}}\cr{}
is the preprint (not reviewed) archivable?
\item \code{postprint}   {\code{character(1)}}\cr{}
is the postprint (reviewed but not formatted) archivable?
\item \code{pdf}         {\code{character(1)}}\cr{}
is the publisher's version (reviewed and formatted)
archivable?
}
}
\description{
Retrieve publisher's policy based on publisher's continent. This function
does not work for unclassified or international publishers.
}
\details{
To provide your API key to \code{rromeo} you can:
\enumerate{
\item provide it as a character string as the \code{key} arguments of \code{rromeo}
functions as \code{rr_*(..., key = "my_key_as_a_string")} (\strong{NOT RECOMMENDED})
\item you can define the variable \code{SHERPAROMEO_KEY} in an \code{.Renviron} file in
your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY=my_key_without_quotes}
\item you can also define the variable \code{SHERPAROMEO_KEY} in an \code{.Rprofile} file
in your working directory, the file should contain the following line
\code{SHERPAROMEO_KEY="my_key_with_quotes"}
}
}
\examples{

rr_publisher_continent(continent = "Caribbean")
rr_publisher_continent(continent = "Central America")
}
