% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{rr_romeo_colour}
\alias{rr_romeo_colour}
\title{Query publisher by RoMEO colour}
\usage{
rr_romeo_colour(romeo_colour = c("green", "blue", "yellow", "white"),
  key = NULL)
}
\arguments{
\item{romeo_colour}{indicates the SHERPA/RoMEO classification of a publisher
see <http://www.sherpa.ac.uk/romeo/definitions.php?la=en&fIDnum=|&mode=simple&version= for definitions of colour>}

\item{key}{a character string or \code{NULL}}
}
\description{
Query publisher by RoMEO colour
}
