% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo-package.R
\docType{package}
\name{rromeo-package}
\alias{rromeo}
\alias{rromeo-package}
\title{rromeo: An R Client for SHERPA/RoMEO API}
\description{
Provides an R Client for SHERPA/RoMEO API
  <http://www.sherpa.ac.uk/romeo/apimanual.php> which indexes policies of
  journal regarding the archival of scientific manuscripts before and/or after
  peer-review as well as formatted manuscripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Rekyt/rromeo}
  \item Report bugs at \url{https://github.com/Rekyt/rromeo/issues}
}

}
\author{
\strong{Maintainer}: Matthias Grenié \email{matthias.grenie@gmail.com} (0000-0002-4659-7522)

Authors:
\itemize{
  \item Hugo Gruson (0000-0002-4094-1476)
}

}
\keyword{internal}
