% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rromeo_functions.R
\name{validate_issn}
\alias{validate_issn}
\title{Checks validity of the ISSN}
\usage{
validate_issn(issn)
}
\arguments{
\item{issn}{The ISSN of a journal}
}
\description{
Silent if the ISSN is valid, errors otherwise.
}
