% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_issn}
\alias{validate_issn}
\title{Checks validity of the ISSN}
\usage{
validate_issn(issn)
}
\arguments{
\item{issn}{\code{\link{character(1)}}\cr{}
The ISSN of a journal}
}
\value{
\code{NULL} if the ISSN is valid, errors otherwise
}
\description{
International Standard Serial Numbers (ISSNs) have a specific structure and
this function checks that the provided ISSN is valid. For more information
please go to
\url{https://en.wikipedia.org/wiki/International_Standard_Serial_Number}
}
\keyword{internal}
